requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.5
pydantic==1.10.12
httpx==0.24.0
anyio==3.7.1

